+++ core/boot/src/test/java/org/eclipse/dataspaceconnector/boot/system/testextensions/ProviderExtension.java
@@ -1,0 +1,26 @@
+/*
+ *  Copyright (c) 2020 - 2022 Microsoft Corporation
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Microsoft Corporation - initial API and implementation
+ *
+ */
+
+package org.eclipse.dataspaceconnector.boot.system.testextensions;
+
+import org.eclipse.dataspaceconnector.boot.system.TestObject;
+import org.eclipse.dataspaceconnector.spi.system.Provider;
+import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
+
+public class ProviderExtension implements ServiceExtension {
+    @Provider
+    public TestObject testObject() {
+        return new TestObject("foobar");
+    }
+}