+++ spi/contract-spi/src/main/java/org/eclipse/dataspaceconnector/spi/contract/negotiation/store/ContractNegotiationStore.java
@@ -17,7 +17,6 @@ package org.eclipse.dataspaceconnector.spi.contract.negotiation.store;
 
 import org.eclipse.dataspaceconnector.spi.persistence.StateEntityStore;
 import org.eclipse.dataspaceconnector.spi.query.QuerySpec;
-import org.eclipse.dataspaceconnector.spi.system.Feature;
 import org.eclipse.dataspaceconnector.spi.types.domain.contract.agreement.ContractAgreement;
 import org.eclipse.dataspaceconnector.spi.types.domain.contract.negotiation.ContractNegotiation;
 import org.jetbrains.annotations.NotNull;
@@ -29,10 +28,7 @@ import java.util.stream.Stream;
  * Stores {@link ContractNegotiation}s and their associated types such as {@link ContractAgreement}s.
  * <p>
  */
-@Feature(ContractNegotiationStore.FEATURE)
 public interface ContractNegotiationStore extends StateEntityStore<ContractNegotiation> {
-
-    String FEATURE = "edc:core:contract:contractnegotiation:store";
 
     /**
      * Finds the contract negotiation for the id or null.