+++ CHANGELOG.md
@@ -12,6 +12,7 @@ in the detailed section referring to by linking pull requests or issues.
 ### Overview
 * Bugfixing DataManagementApi
 * Build improvements
+* Improvements to Dependency Resolution
 
 ### Detailed Changes
 
@@ -22,6 +23,7 @@ in the detailed section referring to by linking pull requests or issues.
 * Check to avoid duplicated module names (#1190)
 * Implement Catalog service for Data Management API (#1195)
 * Add strict body validation for REST endpoints (#1128)
+* Dependency injection using factory/provider methods (#1056)
 
 #### Changed